Fix "No handler set" error from Guzzle's HandlerStack
The "setClass()" method from the decorator overrides the service
definition (which uses a factory, not a class), and caused the compiler
to instantiate the HandlerStack using the following code:

  $a = new \GuzzleHttp\HandlerStack();

instead of:

  $a = \GuzzleHttp\HandlerStack::create();

This caused the default handler stacks from Guzzle not to be added, and
as a result it was raising an exception when trying to use the HTTP
Client.

This was fixed by replacing setClass() with setFactory(), which doesn't
override the parent service definition but also works around the null
class name problem mentioned in d1a46db.

Involved versions:
 - Guzzle 6.1
 - Symfony 2.7.6